Encapsulating withdraw and making funds to be withdrawed right away
