+++ mobile-app/components/WithdrawFundsModal.tsx
@@ -4,16 +4,16 @@ import { useUser } from "@clerk/clerk-expo";
 import { BlurView } from 'expo-blur';
 import { LinearGradient } from 'expo-linear-gradient';
 import { Ionicons } from '@expo/vector-icons';
+import { fetchAPI } from '@/lib/fetch';
 interface WithDrawFundsModalProps {
   visible: boolean;
   onClose: () => void;
-  onWithdrawSuccess: (amount: number) => void; 
   availableBalance?: number; // display max balance
 }
 
 const { height } = Dimensions.get('window');
 
-const WithDrawFundsModal: React.FC<WithDrawFundsModalProps> = ({ visible, onClose, onWithdrawSuccess, availableBalance, }) => {
+const WithDrawFundsModal: React.FC<WithDrawFundsModalProps> = ({ visible, onClose, availableBalance, }) => {
   const [inputAmount, setInputAmount] = useState('');
   const { user } = useUser();
   const animatedValue = useRef(new Animated.Value(0)).current;
@@ -34,6 +34,86 @@ const WithDrawFundsModal: React.FC<WithDrawFundsModalProps> = ({ visible, onClos
     }
   }, [visible]);
 
+  const onWithdrawSuccess = async (withdrawAmount: string) => {
+    try {
+      const parsedAmount = parseFloat(withdrawAmount);
+      if (!withdrawAmount || isNaN(parsedAmount) || parsedAmount <= 0) {
+        return Alert.alert("Invalid amount", "Please enter a valid withdrawal amount.");
+      }
+  
+      // Get the user's Stripe connected account ID
+      const accountResponse = await fetchAPI(`/api/stripe/connected-account?clerkId=${user?.id}`, {
+        method: "GET",
+      });
+  
+      const connectedAccountId = accountResponse.connected_account_id;
+  
+      if (!connectedAccountId) {
+        return Alert.alert("Error", "No connected account found.");
+      }
+      
+      console.log("Above payout fetch")
+      // Attempt payout via your backend
+      const response = await fetchAPI("/api/stripe/payout", {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify({
+          clerkId: user?.id,
+          amount: parsedAmount,
+        }),
+      });
+      
+      console.log("After payout fetch")
+      if (response.needs_bank_account) {
+        return Alert.alert(
+          "Bank Account Required",
+          "You need to add a bank account before withdrawing. Please go to your payment settings to add one."
+        );
+      }
+      console.log("Past bankout")
+      if (response.transfer) {
+        // Update balance in your database
+        const balanceResponse = await fetchAPI("/api/database/balance", {
+          method: "POST",
+          headers: {
+            "Content-Type": "application/json",
+          },
+          body: JSON.stringify({
+            clerkId: user?.id,
+            type: "subtract",
+            amount: parsedAmount,
+          }),
+        });
+        console.log("Past balance")
+        if (balanceResponse.balance) {
+  
+          // Store withdrawal transaction
+          await fetchAPI("/api/database/transactions", {
+            method: "POST",
+            headers: {
+              "Content-Type": "application/json",
+            },
+            body: JSON.stringify({
+              clerkId: user?.id,
+              type: "withdraw",
+              amount: parsedAmount,
+            }),
+          });
+  
+          onClose();
+          Alert.alert("Success", "Withdrawal processed successfully.");
+        }
+      } else {
+        Alert.alert("Error", "Withdrawal failed. Please try again.");
+      }
+    } catch (error) {
+      console.error("Error processing withdrawal:", error);
+      Alert.alert("Error", "Something went wrong. Please try again later.");
+    }
+  };
+  
   const handleConfirmWithdraw = async () => {
     const parsedAmount = parseFloat(inputAmount);
 
@@ -45,7 +125,7 @@ const WithDrawFundsModal: React.FC<WithDrawFundsModalProps> = ({ visible, onClos
       return Alert.alert('Insufficient Balance', 'You cannot withdraw more than your available balance.');
     }
 
-    onWithdrawSuccess(parsedAmount);
+    onWithdrawSuccess(parsedAmount.toString());
   };
 
   return (
